refactor(jsx): extract shared createTemplateAwareStringProtector
Deduplicate splitTemplateInterpolations and the template-literal +
quoted-string protection logic that was copy-pasted between
prop-rewrite.ts (Phase 1) and emit-reactive.ts (Phase 2).

Both callers now use createTemplateAwareStringProtector from
html-template.ts, so future fixes to string protection apply in
one place.
